Update characterization schema
I have revised the characterization schema to have more constraints. It did not result in any issue with both Esen's and my datasets.
Actually the structure of this schema looks a very weird for me, but I did not change it because the conversion scripts probably depend on it.
I have also added an attribute "components" as a place to input the PXRD analysis results from Frederic's script. Ideally it should be located separate from "Characterization" attributes, but I found the "_purity" attribute was already there so I had to put this there.